Retry removing imports cache directory
This could occur if multiple processes are managing a cache at the
same path so simply run fs.removeSync again when an ENOENT error is
thrown will removing the first time.

+++ src/less-cache.coffee
@@ -69,7 +69,13 @@ class LessCache
       if @fallbackDir
         @importsFallbackDir = join(@fallbackDir, basename(@importsCacheDir))
     else if filesChanged
-      fs.removeSync(@importsCacheDir)
+      try
+        fs.removeSync(@importsCacheDir)
+      catch error
+        if error?.code is 'ENOENT'
+          fs.removeSync(@importsCacheDir) # Retry once
+        else
+          throw error
 
     fs.makeTreeSync(@importsCacheDir)
     @writeJson(join(@importsCacheDir, 'imports.json'), {importedFiles})
